CLAMAV-6 Be more strict with looking in form fields for file uploads.
1. File should only belong in Lucee tmp directory
2. File path should not contain any directory traversal

This prevents scary looking errors being thrown in the scanner when an
attacker makes a POST request with a param such as:

`../../../etc/passwd`

This only errors when Lucee will expand that path and the path exists. The error is
then a file not found when we try to pass that file using Java to our scanner
(file not found, because we do not expand it). So it is not a vulnerability, just
very scary looking.